% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_jin_document.R
\name{build_jin_document}
\alias{build_jin_document}
\title{Build the Justice in Numbers Pocketbook documents}
\usage{
build_jin_document(
  doc_type,
  rootpath = "https://data.justice.gov.uk",
  ext = "",
  targetpath = "alpha-jin-pocketbook",
  S3target = TRUE,
  change_check = FALSE
)
}
\arguments{
\item{doc_type}{Either "pocketbook" or "summary_tables" depending on which document you want to create.}

\item{rootpath}{The root URL path for the API data (default is "https://data.justice.gov.uk").}

\item{ext}{Additional extension to be appended to the API file name (default is an empty string).}

\item{targetpath}{The target path for saving the generated document (default is "alpha-jin-pocketbook").}

\item{S3target}{Logical. If TRUE, the generated pocketbook will be saved to S3 (default is TRUE).}

\item{change_check}{Logical. If TRUE, the function will check whether the generated document has changed compared to the latest version on S3 and update accordingly (default is FALSE).}
}
\description{
This function builds the Justice in Numbers Pocketbook or Summary Tables document.
It takes several parameters to customize the document generation process.
}
\examples{
# Generate the pocketbook and save it to the default location in S3 without checking for changes
build_jin_document("pocketbook")

# Generate the pocketbook but only save it to the default location in S3 if it has changed since it was last run
build_jin_document("pocketbook", change_check = TRUE)

# Generate the summary_tables and save them to a folder in the working directory called 'outputs'
build_jin_document("summary_tables", targetpath = "outputs", S3target = FALSE)
}
