% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_summary_tables.R
\name{build_summary_tables}
\alias{build_summary_tables}
\title{Build the Justice in Numbers Summary Tables document}
\usage{
build_summary_tables(
  rootpath = "https://data.justice.gov.uk",
  ext = "",
  targetpath = "alpha-jin-pocketbook/Summary",
  S3target = TRUE,
  change_check = FALSE
)
}
\arguments{
\item{rootpath}{The root URL path for the API data (default is "https://data.justice.gov.uk").}

\item{ext}{Additional extension to be appended to the target file name (default is an empty string).}

\item{targetpath}{The target path for saving the generated document (default is "alpha-jin-pocketbook/Summary").}

\item{S3target}{Logical. If TRUE, the generated document will be saved to S3 (default is TRUE).}

\item{change_check}{Logical. If TRUE, the function will check whether the generated document has changed compared to the latest version on S3 and update accordingly (default is FALSE).}
}
\description{
This function builds the Justice in Numbers Summary Tables.
It takes several parameters to customize the generation process.
}
\examples{
# Generate the document and save it to the default location in S3 without checking for changes
build_summary_tables()

# Generate the document but only save it to the default location in S3 if it has changed since it was last run
build_summary_tables(change_check = TRUE)

# Generate the summary tables document and save it to a folder in the working directory called 'outputs'
build_summary_tables(targetpath = "outputs", S3target = FALSE)
}
