% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_flowchart.R
\name{cjs_flowchart}
\alias{cjs_flowchart}
\title{Build the Justice in Numbers Pocketbook CJS Flowchart page}
\usage{
cjs_flowchart(doc)
}
\arguments{
\item{doc}{A Word document object to which the flowchart page will be added.}
}
\value{
The modified doc object with the CJS flowchart page added.
}
\description{
This internal function builds the Justice in Numbers CJS Flowchart Page.
}
\examples{
# This function is not intended to be called directly by the user.
}
