% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{contents}
\alias{contents}
\title{Build the Justice in Numbers Pocketbook Contents Page}
\usage{
contents(doc)
}
\arguments{
\item{doc}{A Word document object created by the officer::read_docx() function.}
}
\value{
The modified doc object with the table of contents added.
}
\description{
This internal function builds the Justice in Numbers Pocketbook Contents Page.
}
\examples{
# This function is not intended to be called directly by the user.
}
