% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_page.R
\name{cover_page}
\alias{cover_page}
\title{Build the Justice in Numbers Pocketbook Cover Page}
\usage{
cover_page(doc)
}
\arguments{
\item{doc}{A Word document object created by the officer::read_docx() function}
}
\value{
The doc object with cover text added
}
\description{
This function builds the Justice in Numbers Pocketbook Cover Page.
}
