% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JiN_measures.R
\name{JiN_measures}
\alias{JiN_measures}
\title{Build the Justice in Numbers Pocketbook Measures pages}
\usage{
JiN_measures(doc, rootpath = "https://data.justice.gov.uk", ext = "")
}
\arguments{
\item{doc}{A Word document object.}

\item{rootpath}{Root path to the JIN API files (default is "https://data.justice.gov.uk". This should only be changed if testing on downloaded JSON files or the location of the API changes).}

\item{ext}{File extension for API files (default is an empty string. This should only be changed if testing on downloaded JSON files or the format of the API changes).}
}
\value{
The modified doc object with summary tables and charts added.
}
\description{
This internal function builds the Justice in Numbers Measures Pages, covering charts and tables for each JiN measure.
}
\examples{
# This function is not intended to be called directly by the user.
}
