% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tables.R
\name{summary_tables}
\alias{summary_tables}
\title{Build the Justice in Numbers Pocketbook Summary Tables pages}
\usage{
summary_tables(doc, rootpath = "https://data.justice.gov.uk", ext = "")
}
\arguments{
\item{doc}{A Word document object}

\item{rootpath}{Root path to the JIN API files (only required if testing on downloaded JSON files)}

\item{ext}{File extenstion for API files (only required if testing on downloaded JSON files)}
}
\value{
The doc object with summary tables added
}
\description{
This function builds the Justice in Numbers Pocketbook Summary Tables Pages.
}
