% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tables_heading.R
\name{summary_tables_heading}
\alias{summary_tables_heading}
\title{Justice in Numbers - Summary Tables Heading}
\usage{
summary_tables_heading(doc)
}
\arguments{
\item{doc}{A Word document object created by the officer::read_docx() function}
}
\value{
The doc object with the summary tables heading added
}
\description{
This function adds the heading for the Summary Tables section in the Justice in Numbers pocketbook.
}
