% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_flowchart.R
\name{cjs_flowchart}
\alias{cjs_flowchart}
\title{Build the Justice in Numbers Pocketbook CJS Flowchart page}
\usage{
cjs_flowchart(doc)
}
\arguments{
\item{doc}{A Word document object}
}
\value{
The doc object with summary tables added
}
\description{
This function builds the Justice in Numbers CJS Flowchart Page.
}
