% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guidance.R
\name{guidance}
\alias{guidance}
\title{Build the Justice in Numbers Pocketbook Guidance Page}
\usage{
guidance(doc)
}
\arguments{
\item{doc}{A Word document object.}
}
\value{
The modified doc object with guidance added.
}
\description{
This internal function builds the Justice in Numbers Pocketbook Guidance Page.
}
\examples{
# This function is not intended to be called directly by the user.
}
